RS/datamodel: Rename schema -> datamodel
